Split tasks per topic
