+++ src/tasks/index.ts
@@ -1,0 +1,4 @@
+import "./deploy"
+import "./wire"
+import "./verify"
+import "./generateProof"
